refactor(desktop): extract X connector transport
